backend/requirements.txt
# Django Core
Django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Authentication
PyJWT==2.8.0
cryptography==41.0.7

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Background Tasks
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0

# AWS S3
boto3==1.34.14
django-storages==1.14.2

# Utilities
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.6.0

# Monitoring
sentry-sdk==1.39.1

# Development
python-decouple==3.8
